@@ -16,12 +16,12 @@
     "classnames": "^2.3.1",
     "date-fns": "^2.28.0",
     "gray-matter": "^4.0.3",
-    "next": "12.1.6",
+    "next": "^14.1.0",
     "next-mdx-remote": "^4.0.3",
     "next-themes": "^0.2.0",
     "nextjs-progressbar": "0.0.14",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
     "reading-time": "^1.5.0"
   },
@@ -29,7 +29,7 @@
     "@types/node": "^18.0.6",
     "autoprefixer": "^10.4.7",
     "eslint": "8.16.0",
-    "eslint-config-next": "12.1.6",
+    "eslint-config-next": "^14.1.0",
     "postcss": "^8.4.14",
     "tailwindcss": "^3.0.24",
     "typescript": "^4.7.4"
